moved ParseEpisodeId to Util
