+++ cmds/next.js
@@ -1,5 +1,6 @@
 const Discord = module.require("discord.js");
 const fetch = require('node-fetch');
+const Util = require("../Util");
 
 module.exports.run = async (gideon, message, args) => {
     let agc = args[0];
@@ -34,39 +35,10 @@ module.exports.run = async (gideon, message, args) => {
         showtitle = "av2020"; 
         thimg = '';
     } 
-    else return message.channel.send(as);  
-
-    function ParseEpisodeId(input) {
-        if (!input) return null;
-        
-        let str = input.toLowerCase();
-        let s = "", e = "";
-        let hit_limiter = false;
-    
-        for (let letter of str) {
-            if (letter == "s") continue;
-    
-            if (letter == "e" || letter == "x") {
-                hit_limiter = true;
-                continue;
-            }
-    
-            if (!(/^\d+$/.test(letter))) continue;
-    
-            if (!hit_limiter) s += letter;
-            else e += letter;
-        }
-    
-        let s_num = Number(s);
-        let e_num = Number(e);
-    
-        if (isNaN(s_num) || isNaN(e_num)) return null;
-    
-        return "S" + (s_num < 10 ? "0" + s_num : s_num) + "E" + (e_num < 10 ? "0" + e_num : e_num);
-    }
+    else return message.channel.send(as);
 
     const body = await fetch(api).then(res => res.json());
-    const fiep = ParseEpisodeId(args[1]);
+    const fiep = Util.ParseEpisodeId(args[1]);
     if (!fiep) return message.channel.send(es);
 
     let filtered = body.filter(x => x.series == showtitle);
